+++ plugin-dev/Source/Sentry/Private/Apple/SentrySubsystemApple.cpp
@@ -27,6 +27,7 @@
 #include "GenericPlatform/GenericPlatformOutputDevices.h"
 #include "HAL/FileManager.h"
 #include "UObject/GarbageCollection.h"
+#include "Utils/SentryLogUtils.h"
 
 void SentrySubsystemApple::InitWithSettings(const USentrySettings* settings, USentryBeforeSendHandler* beforeSendHandler, USentryTraceSampler* traceSampler)
 {
@@ -200,12 +201,19 @@ TSharedPtr<ISentryId> SentrySubsystemApple::CaptureAssertion(const FString& type
 #elif PLATFORM_IOS
 	int32 framesToSkip = 5;
 #endif
+
+	SentryLogUtils::LogStackTrace(*message, ELogVerbosity::Error, framesToSkip);
+
 	return CaptureException(type, message, framesToSkip);
 }
 
 TSharedPtr<ISentryId> SentrySubsystemApple::CaptureEnsure(const FString& type, const FString& message)
 {
-	return CaptureException(type, message, 6);
+	int32 framesToSkip = 6;
+
+	SentryLogUtils::LogStackTrace(*message, ELogVerbosity::Error, framesToSkip);
+
+	return CaptureException(type, message, framesToSkip);
 }
 
 void SentrySubsystemApple::CaptureUserFeedback(TSharedPtr<ISentryUserFeedback> userFeedback)