+++ plugin-dev/Source/Sentry/Private/SentryOutputDevice.cpp
@@ -11,7 +11,7 @@
 void FSentryOutputDevice::Serialize(const TCHAR* V, ELogVerbosity::Type Verbosity, const FName& Category)
 {
 	const FString Message = FString(V).TrimStartAndEnd();
-	if (Message.IsEmpty())
+	if (Message.IsEmpty() || Message.StartsWith(TEXT("[Callstack]")))
 	{
 		return;
 	}
@@ -74,7 +74,7 @@ bool FSentryOutputDevice::CanBeUsedOnMultipleThreads() const
 	return true;
 }
 
-#if ENGINE_MAJOR_VERSION >= 5
+#if ENGINE_MAJOR_VERSION == 5 && ENGINE_MINOR_VERSION >= 1
 bool FSentryOutputDevice::CanBeUsedOnPanicThread() const
 {
 	return true;