+++ plugin-dev/Source/SentryEditor/Private/SentrySettingsCustomization.cpp
@@ -17,7 +17,7 @@
 #include "PropertyHandle.h"
 #include "Framework/Notifications/NotificationManager.h"
 #include "HAL/FileManager.h"
-#include "HAL/PlatformFileManager.h"
+
 #include "Interfaces/IPluginManager.h"
 #include "Runtime/Launch/Resources/Version.h"
 
@@ -33,6 +33,12 @@
 #include "Styling/AppStyle.h"
 #else
 #include "EditorStyleSet.h"
+#endif
+
+#if ENGINE_MAJOR_VERSION >= 5
+#include "HAL/PlatformFileManager.h"
+#else
+#include "HAL/PlatformFilemanager.h"
 #endif
 
 const FString FSentrySettingsCustomization::DefaultCrcEndpoint = TEXT("https://datarouter.ol.epicgames.com/datarouter/api/v1/public/data");
@@ -411,8 +417,16 @@ TSharedRef<SWidget> FSentrySettingsCustomization::MakeLinuxBinariesStatusRow(FNa
 
 					FString CommandLine = FString::Printf(TEXT("BuildPlugin -Plugin=\"%s/Sentry.uplugin\" -Package=\"%s\" -CreateSubFolder -TargetPlatforms=Linux"), *PluginPath, *TempLinuxBinariesPath);
 
-					IUATHelperModule::Get().CreateUatTask(CommandLine, FText::FromString("Windows"), FText::FromString("Compiling Sentry for Linux"), FText::FromString("Compile Sentry Linux"),
-						FAppStyle::GetBrush(TEXT("MainFrame.CookContent")), nullptr, [this, TempLinuxBinariesPath](FString result, double X)
+					IUATHelperModule::Get().CreateUatTask(CommandLine, FText::FromString("Windows"),
+						FText::FromString("Compiling Sentry for Linux"),
+						FText::FromString("Compile Sentry Linux"),
+					#if ENGINE_MAJOR_VERSION == 5 && ENGINE_MINOR_VERSION >= 1
+						FAppStyle::GetBrush(TEXT("MainFrame.CookContent")),
+						nullptr,
+					#else
+						FEditorStyle::GetBrush(TEXT("MainFrame.CookContent")),
+					#endif
+						[this, TempLinuxBinariesPath](FString result, double X)
 						{
 							if (result.Equals(TEXT("Completed")))
 							{