+++ scripts/download-sdks.sh
@@ -23,7 +23,7 @@ if [[ "$runId" == "" ]]; then
     exit 1
 fi
 
-declare -a sdks=("Android" "IOS" "Linux" "Mac")
+declare -a sdks=("Android" "IOS" "Linux" "LinuxArm64" "Mac")
 for sdk in "${sdks[@]}"; do
     echo "Downloading $sdk SDK to $PWD/$sdk ..."
     rm -rf "./$sdk"